Fixing some 64 bit warnings in function calls.

+++ src/rumble.h
@@ -490,7 +490,7 @@ typedef struct
         const char  *port;
         const char  *name;
         int         threadCount;
-        size_t      stackSize;
+        unsigned int stackSize;
     } settings;
 } rumbleService;
 typedef struct
@@ -646,7 +646,7 @@ char                        *rumble_sha160(const unsigned char *d); /* SHA1 dige
 char                        *rumble_sha256(const unsigned char *d); /* SHA-256 digest (64 byte hex string) */
 char                        *rumble_decode_base64(const char *src);
 char                        *rumble_encode_base64(const char *src, size_t len);
-int                         rumble_unbase64(unsigned char *dest, const unsigned char *src, int srclen);
+int                         rumble_unbase64(unsigned char *dest, const unsigned char *src, size_t srclen);
 void                        rumble_string_lower(char *d);           /* Converts <d> into lowercase. */
 void                        rumble_string_upper(char *d);           /* Converts <d> into uppercase. */
 rumble_args                 *rumble_read_words(const char *d);
